Add files via upload
The dataset has information for weekly sales (taken each Friday) for 45 stores from Feb 2010 till October 2012. 
We have 635 rows and 8 columns. and The columns include the store number, the date of the sale, the weekly sales amount, whether the sale occurred during a holiday period or not, the temperature, fuel price, CPI (Consumer Price Index), unemployment rate, and the holiday name.
We wanted to see how, from the perspective of Walmart, the stores could be managed in the best possible way.